ggml: refactor fp16<->fp32 simd to ggml-cpu
